[all] Remove warnings
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@5083 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/common/ADM_editor/ADMedAVIAUD.cpp
@@ -109,7 +109,7 @@ again:
             printf("[Composer::getPCMPacket] drift %d, computed :%lu got %lu\n",(int)(lastDts-packetBufferDts),lastDts,packetBufferDts);
             if(packetBufferDts<lastDts)
             {
-                printf("[Composer::getPCMPacket] Dropping packet %"LU" last =%"LU"\n",lastDts/1000,packetBufferDts/1000);
+                printf("[Composer::getPCMPacket] Dropping packet %"LU" last =%"LU"\n",(uint32_t)(lastDts/1000),(uint32_t)(packetBufferDts/1000));
                 drop=true;
             }else 
             {
